backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
pydantic-settings==2.4.0
sqlalchemy==2.0.35
sqlmodel==0.0.21
anthropic==0.34.2
sentence-transformers==3.1.1
numpy==1.26.4
PyYAML==6.0.2
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
openai==1.40.0
